Add evaluating functions and new variables
Added lstOperator(array), total(int) and evaluate(boolean). Might use evaluate
in future.

When operator is clicked on calculator, it will be appended to lstOperator.
If lstOperator length == 1, the next nums clicked will be appended
to secondOperand. Otherwise, the nums will be appended to firstOperand.

If lstOperator > 1, then one of the evaluating functions will be called on.
(will check first item of lstOperator) ie/ ['+', '-'] --> add()

When equation is executed, the operator will be removed with .pop function
from lstOperator.